Create new folders and rearrange source files
